backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.2

# Image processing
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
scikit-image==0.24.0

# OCR
paddleocr==2.7.3
paddlepaddle==2.6.1

# PPT
python-pptx==1.0.2
lxml==5.3.0
